python-service/requirements.txt
flask
flask-cors
PyMuPDF
google-generativeai
python-dotenv
